objects marked in SPARK have types marked in SPARK
Asserting that objects marked in SPARK have their types marked in SPARK
seems obvious, but it wasn't true on QB07-050__single_po, where we have:

   protected PO is
   private
      C : Integer := PO'Size;
   end;

Here marking thinks that in "PO'Size" the attribute is applied to object,
while in fact it applies to type. This patch fixes this problem: once we
detect that the default expression refers to the current instance of a
protected type, we don't mark it.

(no-tn-check)

* gnat2why/spark/spark_definition.adb
(Mark_Object_Declaration): add a seemingly obvious assertion.
(Mark_Default_Expression): do not mark expressions that refer to the current
 type instance.

* testsuite/gnatprove/tests/O110-002__default_init/test.out
Update output; violation in type T is now correctly propagated.

Change-Id: Ica0688514aa99d6b418ca69e4a0b6402026dff29