S415-012 handling of pledge functions to annotate local borrowers
Functions annotated with a Pledge annotate pragma are translated in a
special way in SPARK. They express that the boolean expression in their
second parameter should hold at the end of the scope of the borrower in
their first parameter, and so, no matter what modifications are done to
either the borrowed entity or the borrower in the mean time. We
roughly translate Fun_Pledge (Borrower, P (Borrower, Borrowed)) as
forall borrower, borrowed.
    pledge (borrower, borrowed) -> p (borrower, borrowed)

* spark_definition.adb
(Ownersip_Checking): Supply value for new parameter Is_Pledge_Function.

* gnat2why/why/why-gen-pointers.adb
(Borrow_Info): Add a component for the borrowed entity.
(Declare_Ada_Pointer): Declare the dummy object which was forgotten.
(Declare_Pledge_Ref): Change the pledge predicate to relate the borrowed
 entity with the borrower and not the borrowed expression. This is so that
 we have better triggers for the quantifier in the translation of calls
 to pledge functions.
(Get_Borrowed_Entity): Accessor for the borrowed entity.

* gnat2why/why/gnat2why-expr.adb
(Compute_Assumption_For_Borrow): Compute the properties on the prefix of
 a borrowed expression which are frozen by a borrow (value of discriminants,
 array bounds...).
(Transform_Pledge_Call): Transform a call to a function annotated with a
 pragma Annotate (GNATprove, Pledge, ...).
(Havoc_Borrowed_Expression): Adapt the havoc program to the new signature
 of pledge predicates.
(Update_Value_From_Assignment): Adapt the assignment statement to the new
 signature of pledge predicates.
(Transform_Expr): Call Transform_Pledge on calls of function annotated with
 a pragma Annotate (GNATprove, Pledge, ...).

* gnat2why/why/gnat2why-types.adb
(Declare_Type_Completion): Add conversion to the expected subtype for
 parameters of a redefined equality function. Apparently this subtype may
 not exactly match the base type.

* testsuite/gnatprove/lib/python/test_support.py
Add marker for dereference checks.

* tests/S415-012__ex_pledge
New test.

* tests/S415-012__iterate_list
New test, this one could be distributed someday.

Change-Id: I7c98ffb5282dd07974834e084a53d9f75ced0cd3