R906-030 generate axiom for recursive functions
We now generate an axiom for the postcondition of recursive functions.
We do not include it when proving the subprogram itself and mutually
recursive subprograms as it would be unsound. We do not include it either
when proving a package elaboration, or VCs for a type declaration as
it is not clear yet how we can decide whether those are mutually recursive
with a function or not.

* gnat2why/flow/flow_generated_globals-phase_2.ads
(Mutually_Recursive): Return True if two subprograms are mutually
 recursive.

* gnat2why/why/gnat2why-expr.adb
(Transform_Function_Call): Do not warn on calls of recursive functions
 anymore, as whether or not the axiom is present depends on the entity
 we are proving. We rather warn on function declarations.

* gnat2why/why/gnat2why-subprograms.adb
(Generate_Axiom_For_Post): For terminating recursive functions, generate
 the post axiom in a distinct theory so that it can be included or not
 separately.

* gnat2why/why/gnat2why-util.adb
(Has_Post_Axiom): Return True for subprograms for which we are going to
 generate a post axiom.

* gnat2why/why/why-atree-modules.ads
(Rec_Axiom_Modules): Map for post axiom modules of recursive functions.
(E_Rec_Axiom_Module): Return the post axiom modules of a recursive
 function.

* gnat2why/why/why-inter.adb
(Add_Axiom_Imports): Add imports for post axiom modules of recursive
 functions.

* include/spark-higher_order-fold.adb
(Count_2): Help to restore proof by providing a (redundant) loop invariant

* tests/R509-011__higher_order_proof
Update session test.

* tests/O512-022__number_theory/test.out
Some proofs are lost, can be regained with more steps.

* tests/L323-005__recursive/test.out
* tests/R906-030__recursion/test.out
* tests/S415-012__ex_pledge/test.out
* tests/S415-012__iterate_list/test.out
* tests/adacore_u/test.out
Small changes in provability.

* tests/L323-006__recursive_expr_func/test.out
* tests/N404-022__mutually_rec_post/test.out
* tests/P725-016__recursive_spec/test.out
* tests/P725-016__recursive_bad_axiom/test.out
* tests/L323-005__recursive/test.out
* tests/K330-014__loop_entry/test.out
* tests/M503-001__typinv/test.out
* tests/N624-020__perm_occ/test.out
* tests/N624-020__perm_rem/test.out
* tests/O512-022__prefix_sum/test.out
* tests/O902-020__private_types/test.out
* tests/PC08-006__flow_terminating_message/test.out
* tests/Q424-003__if_expression/test.out
* tests/Q615-017__factorial/test.out
* tests/Q802-009__count_locations/test.out
* tests/QA11-003__itp_sumrange/test.out
* tests/R728-002__aggregates/test.out
* tests/S325-002__anonymous_access/test.out
* tests/S515-024__pointer_traversal/test.out
* tests/ghc_sort/test.out
* tests/pair_insertion_sort/test.out
Changes in info messages only.

Change-Id: Iecd89735c0debfe4f459a019d563e2dcbdda5571