S702-024 record attributes on dereferences
Fix value of the record attributes 'Tag and 'Constrained when they are
applied to a pointer dereference. On allocators, 'Tag is taken either
from the designated type if it is specific, or from the input value if
it is classwide. 'Constrained is always true as allocated values are
constrained by their initial value.
Additionnaly, we had to adapt the handling of the 'Constrained attribute
so that it returns true if its prefix denotes a constant, as dereferences
inside a constant object are not themselves constants.

* gnat2why/why/spark_atree.adb
(Get_Enclosing_Object): Stop using Sem_Util.Get_Enclosing_Object. The new
 version returns the enclosing object even if there is a dereference in the
 path.

* gnat2why/spark/spark_util.adb
(Has_Dereferences): Return True if an expression contains dereferences
 on its biggest suffix which is a path. So it will return True on Expr.all.C
 but not on T'(Expr.all.C).

* gnat2why/why/gnat2why-expr.adb
(Transform_Attr): 'Constrained on a part of an expression which is not a
 path or on a part of a constant will return True, but only if the part does
 not go through a pointer dereference.
(Transform_Expr): Assume correct values for 'Tag and 'Constrained attributes
 on allocated values. 'Constrained is always true for allocated values and
 'Tag either comes from the designated type if it is specific, or from the
 initial value if it is not.

* gnat2why/flow/flow_error_messages.adb
(Get_Explanation): Use Get_Enclosing_Object from SPARK_Atree to go through
 pointer dereference.

* gnat2why/why/why-gen-records.adb
(New_Record_Attributes_Update): Change the behavior when both From_Expr and
 Is_Cst are supplied. It does not change much as the function is never
 called that way, but the behavior is closer to its description in the ads.

* tests/S702-024__record_attributes_in_allocators
New test for soundness of 'Constrained and 'Tag on pointers.

Change-Id: I22960794bc8a32d9b6448dd80ccd75ef9de6a69e