RC21-009 Fix test
RC03-030__conf_file is a test for adding a prover via the why3config tool
and testing that it works in SPARK.
This is the completion of this test. Note that this is important that the
version of the prover is recognized by why3config
(why3/share/prover_detection_data.conf) in order for this to work. So, I
simulated an outdated version of Z3 which hopefully is not present on dev
setup (if it is it would not work).
The idea is to add this prover in an external config built in the test
(toto.conf) and then check that this works by checking that it proves when
called a first time. Then we remove the script file and check that it
fails (so that we are sure that the script was used).
The script is the simplest we could do.

* testsuite/gnatprove/tests/RC03-030__conf_file/test.opt
Removed

* testsuite/gnatprove/tests/RC03-030__conf_file/test.out
Updated.

* testsuite/gnatprove/tests/RC03-030__conf_file/test.py
Add the path to current directory into the PATH variable
Change executable name so that it is recognized by why3config (it seems
necessary).

* testsuite/gnatprove/tests/RC03-030__conf_file/z3-4.4.0
Compiled from src/main.adb (compiled during the test and not present in
git).

* testsuite/gnatprove/tests/RC03-030__conf_file/src*
Small compilable ada code used to generate the fake executable for Z3.
When -version, output the standard version of Z3 output (parsable by
why3config)
Else, answer proved (ie "unsat").

Change-Id: Ic5bf140d41cbb1a00e374fd85e77220daab03b2a