S708-001 Use fakecvc3 instead of fakez3
This patches test RC03-030__conf_file: z3 cannot be added as an external
prover.

* testsuite/gnatprove/tests/RC03-030__conf_file/src/main.adb
Now behaves like a cvc3.

* testsuite/gnatprove/tests/RC03-030__conf_file/test.py
Do PATH manipulation so that the configuration file contains *only* the
fakecvc3 that we built (otherwise if a developper has a cvc3 on his
machine the test would be broken for her).
Replace all occurences of z3 with cvc3.

* testsuite/gnatprove/tests/RC03-030__conf_file/test.out
Update output

Change-Id: I6cef37dd2174136b1434a11d4ef8bec36a925143