S711-064 Fix handling of actions in shorcut expressions
The semantics of actions in shortcut expressions was imprecisely modeled, so
that run-time checks were unprovable. The precise semantics is that such
actions need only be evaluated when the rhs of the shortcut expression
(and-then, or-else) is evaluated. This allows GNATprove to be more precise.

* gnat2why/why/gnat2why-expr.adb
(Transform_Expr): Change handling of actions in short circuits, so that
 they are only evaluated when the righ-hand side is evaluated.

* testsuite/gnatprove/tests/S711-064__pointer_to_array/test.opt
* testsuite/gnatprove/tests/S711-064__pointer_to_array/test.out
* testsuite/gnatprove/tests/S711-064__pointer_to_array/test2.ads
Test passes now. Add test without pointers.

Change-Id: I962b2676fa9fd60473f771423dc16c34a6157692