S705-011 special-case formal IN of access types
Formal parameters of mode IN with a non-constant access type can be
assigned, so we handle them very much like IN OUTs; except that we suppress
a warning when they are not actually written.

* gnat2why/flow/flow-control_flow_graph.adb
(Create_Initial_And_Final_Vertices): special-case formal IN parameters of
 non-constant access types.

* gnat2why/flow/flow-analysis.adb
(Find_Unwritten_Exports): special-case formal IN parameters of named access
 types.

* testsuite/gnatprove/tests/RA12-023__pointers/test.out
* testsuite/gnatprove/tests/RA12-026__pointers/test.out
* testsuite/gnatprove/tests/S325-002__anonymous_access/test.out
* testsuite/gnatprove/tests/S415-012__ex_pledge/test.out
* testsuite/gnatprove/tests/S415-012__iterate_list/test.out
* testsuite/gnatprove/tests/S503-019__flow_detect_borrow_aliasing/test.out
* testsuite/gnatprove/tests/S705-011__flow_pointer/test.out
Spurious warnings about unused assignments to pointer IN formal parameters
and about such assignments having no effect are gone. New warnings that
suggest adding a "constant" qualifier to an anonymous access type to an IN
formal parameter.

Change-Id: I9029a883adc90228964270547448dbffb561f557