S523-015 fix sanity checking of user-defined equality on RECORD types
Raise a error message in flow if the Global aspect is not null for
user-defined equality of record types, as per SPARK RM 6.6(1).

* gnat2why/flow/flow-analysis-sanity.adb
(Check_Expressions): If the Global aspect is not null for user-defined
equality of record types, raise an error message, thus aligning with
SPARK RM 6.6(1).

* testsuite/gnatprove/tests/Q308-009__primitive_eq/test.out
* testsuite/gnatprove/tests/S523-015__flow_eq/test.out
Update test outputs.

Change-Id: I2c99c6d057051851a92be51d25910a93d719d008