S617-022 support local borrowers in flow
We now keep track of local borrowers very much like it happens in both the
borrow checker and in proof. When borrowing, we copy the borrowed object
into the borrower; when returning, we do the opposite. For this we keep the
borrowed & borrowed objects on a stack.

* gnat2why/flow/flow-control_flow_graph.adb
(Context): add component to track local borrowers.

(Do_Subprogram_Or_Block): keep track of local borrowers and pop them from
 the stack before leaving; adapt pretty-printing of the CFG as needed.

(Do_Object_Declaration): detect declarations of local borrowers and put them
 on the stack.

Change-Id: I39d2b1f4541581bd665191ad9c21808cb9f327b6