S721-004 add minimal reproducer for crash due to inlining for proof
This is a minimal reproducer based on the GNAT's Sem_Cat unit. It crashes
when flow sees a constant while traversing actions. This constant comes from
inlining for proof, but was not flagged as internal (probably by mistake).
However, we flow should not analyse complex actions in the first place.

* testsuite/gnatprove/tests/S721-004__inline/
Add; crashes, so disabled.

Change-Id: Ibabc73e79143a9185d4deb0c06f2c33df4c7e8a5