S725-017 fix range check on booleans
A conversion with range check from Bool to Bool has to go through
int, because Boolean.range_check takes ints. This requires doing a
conversion in a bit less cases *before* the range check, and in a
bit more cases *after* the range check.

* why-gen-expr.adb
(Insert_Scalar_Conversion): do not convert a boolean to Bool before
the range check, but after

Change-Id: I34bdf9399f59b8f5662a106dc3e62a007f83cdc3