S726-005 fix failing sanity checks in Component_Is_Visible_In_SPARK
In Component_Is_Visible_In_SPARK, we try to decide whether a component of
a record should be visible from the SPARK only part of a program. This
function is called from marking, so it should work on partly marked nodes.

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): Do not attempt further marking on a component whose
 type is not in SPARK to avoid having to deal with non-SPARK types in
 routines dealing with visibility of record components.

* gnat2why/spark/spark_util.adb
(Component_Is_Visible_In_SPARK): The type in which the component is
 declared should have been marked already. This type is either the scope
 of the component or its full view if the type is private. We cannot
 assume that the private view has been marked, so we should not use
 routines which only work on marked type on it.

Change-Id: I294a91b14b4921b86ba670050574d62eac866543