S723-008 Do not create VCs that are not passed on to gnatwhy3
The code for creating the mapping for prefixes of 'Old should
not do that as a side-effect of creating the corresponding Why3 programs
for the postcondition and contract cases, as this leads to VCs being
created and dropped. This is bad for parallel proof of programs.
Instead, postcondition and contract cases are traversed and the mapping
is updated for every occurrence of 'Old.

* gnat2why/why/gnat2why-subprograms.adb
(Generate_VCs_For_Subprogram): Do an explicit traversal to build the map.

Change-Id: Idb604155c915ef9fc1e14bfe146dfa95bf2683e7