S720-001 Fix translation of choice in case for static predicates
When a choice in a case-expression or case-statement is a type with
static predicate, the full range of the (discrete) type should not be
used. Instead, the values implied by the predicate should be used.

* gnat2why/spark/spark_util-types.ads
(Has_Static_Predicate, Static_Discrete_Predicate): Import functions
 from Einfo.

* gnat2why/why/gnat2why-expr.adb
* gnat2why/why/gnat2why-expr.ads
(Transform_Discrete_Choice): Add special case when a static predicate
 applies.

* testsuite/gnatprove/tests/S720-001__static_predicate/lexer.adb
* testsuite/gnatprove/tests/S720-001__static_predicate/lexer.ads
* testsuite/gnatprove/tests/S720-001__static_predicate/predicates.ads
* testsuite/gnatprove/tests/S720-001__static_predicate/test.out
* testsuite/gnatprove/tests/S720-001__static_predicate/test.py
Add test.

* testsuite/gnatprove/tests/sparksmt__lexer_termination/test.out
* testsuite/gnatprove/tests/sparksmt__lexer_termination/test.py
Update output, as loop variant is not provable in fact.

Change-Id: I33fa8ce87aa3371c1a82045c4fd9a3a1e5e5c7ba