S724-033 Add information message for contextual analysis
When subprogram are only analyzed in the context of calls, using
switch --info now displays the following information to users:

  info: local subprogram "Foo" only analyzed in the context of calls
  info: add a contract to analyze it separately from calling contexts

This is useful to understand why some subprograms are not analyzed,
when the call sites are all outside of the SPARK code.

* docs/ug/en/source/how_to_write_subprogram_contracts.rst
Refine the explanation on messages about contextual analysis.

* gnat2why/spark/spark_util-subprograms.adb
* gnat2why/spark/spark_util-subprograms.ads
(Analysis_Requested): Add a variant of the function that distinguishes
 beyond multiple reasons for not analyzing a subprogram.

* gnat2why/why/gnat2why-driver.adb
(Generate_VCs): Issue a message when --info is set and a subprogram
 is only contextually analyzed.

* testsuite/gnatprove/tests/*/test.out
Update outputs.

* testsuite/gnatprove/tests/S724-033__inlining/test.adb
* testsuite/gnatprove/tests/S724-033__inlining/test.ads
* testsuite/gnatprove/tests/S724-033__inlining/test.gpr
* testsuite/gnatprove/tests/S724-033__inlining/test.out
* testsuite/gnatprove/tests/S724-033__inlining/test.py
Add test.

Change-Id: I9b5990ef7082765c51591b0ccdb0a6378dfb1e74