S725-044 more precise dependency tracking for protected operations
Protected operations have an implicit formal parameter, i.e. the enclosing
protected type; when those operations are called, we need to find an actual
that corresponds to the implicit parameter. For calls to standalone
protected objects, e.g.:

    Protected_Object.Protected_Proc;

we simply used the entire "Protected_Object". However, when the protected
object was a component of a record, e.g.:

    Record.Protected_Object.Protected_Proc;

we also used the entire object, i.e. "Record", which was less precise than
we could. Now we take "Record.Protected_Object".

* gnat2why/flow/flow_utility.ads
* gnat2why/flow/flow_utility.adb
(Get_Protected_Object): new routine to get the most precise object; use it
 roughly at all places where we make use of Is_External_Call.

* gnat2why/flow/flow-interprocedural.adb
(Add_Simple_Procedure_Dependency): flip conjuncts of a condition to avoid
 call to Is_Implicit_Input, which would crash on record components. Now we
 will never call it because the implicit actual parameter is always
 contained in the Globals.Inputs set (i.e. it is always an input of a
 protected operation).

* testsuite/gnatprove/tests/S725-044__flow_protected_depends/test.out
Update output; spurious check is gone.

Change-Id: I9233a3e81e506aa2d1e95f570e2e56fa44922eb4