simplify handling of constants that capture values in actions
In Get_Variables we had a post-processing step for constants introduced by
the frontend when processing actions (which is related to expressions with
potential side-effects). We replaced those constants with objects referenced
in their initialization expressions. Now we do this immediately when
discovering such a constant, which is easier and more consistent with a
handling of internal variables.

Also, the new code feels more efficient, though now we traverse the AST
several times, i.e. each time we see a reference to the internal constant.
However, we don't expect many such references and at least now we don't
iterate over the entire container in the post-processing step.

(no-tn-check)

* gnat2why/flow/flow_utility.adb
(Get_Variables): replace post-processing with immediate recursing.

Change-Id: Ia8d1f05b62cd6db30e23eba419164e93b39646d5