S730-064 simplify detection of out-of-memory error in OCaml
Previously we detected out-of-memory string on the normal execution path,
i.e. even when JSON parsing succeeded, so we could argue that performance
is important. Now we only detect this string on the exceptional path, i.e.
when a parsing error certainly occurred, so the performance doesn't really
matters.

* gnat2why/why/gnat2why-error_messages.adb
(Parse_Why3_Prove_Result): rewrite to make this code more readable.

Change-Id: Ia176010232428b9f8a5db00d98b50d5511d4e574