S801-023 relocate DIC check with no reference to current type instance
When a DIC does not have references to the current type instance, it is
considered a precondition of the default initialization code. Thus, it
is checked when variables of the type are default initialized, and not
at the private type declaration.
When variables with such a DIC are default initialized, we now
locate the DIC check on the variable declaration, and not on the DIC, to
facilitate debugging of user code. We don't do the same thing for
record components for now, as we have not decided what the best location
would be. DIC of record components are located on the component declaration.
Also relocate predicate check on default value of record components on the
component rather than on the default initialized variable.

* gnat2why/why/gnat2why-expr.adb
(Compute_Default_Check): Use the Ada_Node to locate DIC checks. Give the
 component entity for the Ada node of recursive calls on record
 components.

* testsuite/gnatprove/tests/S801-023__false_DIC
Customer test.

* testsuite/gnatprove/tests/P127-017__two_wrong_dic/test.out
* testsuite/gnatprove/tests/PA12-001__DIC/test.out
* testsuite/gnatprove/tests/Q512-013__flow_type_aspects/test.out
Update output with change of DIC check locations.

* tests/P118-002__predicate_component/test.out
* tests/P301-005__pred_rec_decl/*
Update output with change of predicate check locations.

Change-Id: Iace8e259e6a0e4062c915523acca107bad6b17b7