S806-027 detect Proof_Ins read in delay statement expression
When detecting references that affect the value of an expression (i.e. when
Get_Variables is called with Fold_Functions => True), we must also detect
references to Proof_Ins in the called functions.

We did this in other statements that evaluate an expression, e.g. in
Do_Case_Statement and in Do_If_Statement; now we also do this in
Do_Delay_Statement.

* gnat2why/flow/flow-control_flow_graph.adb
(Do_Delay_Statement): append delay statement expression to the list of nodes
 where Proof_Ins must be detected.

* testsuite/gnatprove/tests/S806-027__flow_delay/
Update output with an expected check.

Change-Id: I4f71e138f71af122fac11376967a2acb25efd1ab