S730-045 defer DIC check referring to discriminants of the type instance
When the Default_Initial_Condition of a private type refers to the current
type instance, it is checked on the type declaration, otherwise it is
checked on default initialized variables of this type.
We now check Default_Initial_Conditions only referring to the type instance
through discriminants on default initialized variables.

* gnat2why/why/gnat2why-util.adb
(Check_DIC_At_Declaration): Moved away to SPARK_Util.Types so that it can
 use frontend queries instead of flow analysis to search for references to
 the current type instance.

* gnat2why/spark/spark_util-types.adb
(Check_DIC_At_Declaration): Moved from GNAT2Why.Util. Traverse the DIC
 expression to find references to the current type instance. Go up the
 AST using Parent to see if it is actually a reference to a discriminant
 or not.

* tests/S730-045__dic_with_discrs
File test.

Change-Id: I79f6416d3a13ff3b98f91e0ef1a84daddac0ec1d