S807-028 rewrite handling of Depends on protected subprograms
Previously we handled the implicit parameter of protected subprograms like a
global. Consequently, when the protected object was a component of an array,
we missed all variables referenced (and subprograms called) in the array
index expression.

Now, for external calls, we handle the implicit parameter more like an
actual; for internal calls we do what we did.

* gnat2why/flow/flow-control_flow_graph-utility.ads
* gnat2why/flow/flow-control_flow_graph-utility.adb
(Make_Implicit_Parameter_Attributes): adapt to new handling.

* gnat2why/flow/flow-control_flow_graph.adb
(Do_Call_Statement): adapt to new handling.

* gnat2why/flow/flow-interprocedural.adb
(Find_Parameter_Vertex): remove code for Record_Field of a global or
 implicit parameter; both are now always represented as Direct_Mapping.

(Add_Simple_Procedure_Dependency): implicit parameter is represented here
 like any other formal parameter.

* gnat2why/flow/flow_utility.ads
* gnat2why/flow/flow_utility.adb
(Get_Depends): remove Callsite parameter; it was always suspicious and it is
 no longer used; adapt callers as needed.

(Do_Subprogram_Call): adapt to new handling.
(Get_Protected_Object): remove; no longer needed.

* testsuite/gnatprove/tests/S807-028__flow_protected_array/test.out
Update; now as expected.

* testsuite/gnatprove/tests/P602-007__protected_no_return/test.out
Update; the new messages are acceptable (though not terribly intuitive),
because this test it intentionally contrived. (To get those messages we
would have to add a No_Return statement to the caller, but this is currently
rejected by the frontend.)

Change-Id: I9c51d9791630f946ebc1cd1321d7848a15a21aa4