(no-tn-check) improve timing code
Allow for a phase to be registered several times. The timings for
all occurences of a phase will be added together.

Mainly we replace the previous list of timings by a map, and need to
query this map when adding a new timing.

* debug-timing.adb
(External_Timing): renamed to Register_Timing
(Register_Timing): if an entry already exists for the timing, add
  the new time to the existing time. Moved debug code here from
  Timing_Phase_Completed.
(Timing_Start): use map instead of list
(Timing_Phase_Completed): move debug code from here to
  Register_Timing; use Register_Timing to register time
(Timing_History): need to query map instead of list

Change-Id: I745707fc67b2687b507602a1449d36abff58c06b