S806-016 Fix coq lemma library
This moves all Coq files *.ctx from subfolder to proof/Coq. The former
organization was used to mimick the naturally generated way gnatprove/Why3
organize Coq/session files.
Now, files are not put under a lot more additional directories. This would
be too cumbersome to organize (also the directory is already in the files)
so the organization into directories is removed.
The same change is done to the Makefile.
This also removes the obsolete session files that are not used anymore.

* include/proof/sessions/spark-*/why3session.xml
Remove now useless sessions

* include/proof/Coq/*/*.ctx -> include/proof/Coq/*.ctx

* include/Makefile
Adapt the locations where *.ctx files should be found by the makefile
(for cpp preprocessor parsing)

Change-Id: I16e9a414f1a2f405d6a53325441cd2113185d349