S430-007 fix crash in spark_memcache_wrapper
When running the P922-044__memcpy_revised_stars2 test with --cache switch we
are getting a big counterexample (~9MB). Such a large value was crashing the
memcache wrapper, because we were creating a temporary string that didn't
fit on the stack.

* common/memcache_client.adb
(Set): avoid concatenation with an arbitrarily long string for the Value.

Change-Id: Ief7ae8560ec4b3d85b2ee85eaa6b9ac68d7a53fb