S818-003 suppress warnings on calls with no effects
Flow contracts for subprograms with no effects, i.e. no formal parameters
and no global parameters, are "Global => null" and "Depends => null" and
skipping any of these contracts should make no difference to the analysis of
the caller.

However, when Depends was missing, an artificial write to a "null export"
were added to CFG (which was considered to be an effect, because this "null
export" rightly has a property of Effective_Writes); if Depends was present,
we didn't get this write and the entire call appeared as having no effect,
thus a warning was emitted.

Now we only add this artificial write to CFG when necessary and we
explicitly suppress warnings about calls to subprograms with no effects.
Instead, we emit a warning when analysing the subprogram with no effects
itself (unless it is annotated with explicit Global/Depends => null).

* gnat2why/flow/flow-analysis.adb
(Is_Dummy_Call): new routine to detect subprograms with no effects.
(Find_Ineffective_Statements): suppress warnings with Is_Dummy_Call.

* gnat2why/flow/flow-control_flow_graph.adb
(Do_Call_Statement): only add "null export" vertex when necessary, i.e.
 when an implicit/formal/global parameter will flows into it.

* gnat2why/flow/flow-interprocedural.adb
(Add_Simple_Procedure_Dependency): reflect change in Do_Call_Statement in
 code that connects inputs to outputs of a subprogram call.

* testsuite/gnatprove/tests/S818-003__flow_null/test.out
Update; dedicated test for this TN, where spurious warnings are gone.

* testsuite/gnatprove/tests/ironsides/test.out
Update; just like above, a spurious warning is gone.

* testsuite/gnatprove/tests/M611-010__no_return_illegal/test.out
* testsuite/gnatprove/tests/O521-002__no_return/test.out
* testsuite/gnatprove/tests/P602-007__protected_no_return/test.out
* testsuite/gnatprove/tests/P715-019__flow_subprogram_termination/test.out
Auxiliary checks are gone. Those checks do not correspond to any rule in the
SPARK RM and were added as a sort of heuristic. Note that they only appear
when using No_Return aspect, especially when combined with generating
globals, which still isn't working great. Anyway, it is harmless for those
checks to disappear.

* testsuite/gnatprove/tests/O914-009__limit_subp_for_entries/test.out
New warning that comes from an unrelated problem with handling of entry
barriers.

Change-Id: I8ac5628a3c7867612399454fa8092a980c03eef1