S901-003 another test with predicate functions
This tests demonstrates that flow detects subprograms that do not terminate
due to evaluation of a non-terminating type predicate expressions.

Note: info message in line 9 about expression function body not available
for proof is kind of spurious, because it refers to an internal expression
function, i.e. a predicate function.

* testsuite/gnatprove/tests/S901-003__flow_term_pred/bad.adb
Add; output is mostly correct.

Change-Id: Iea63362c91fa2cb155384db713f60a1f2ff61413