S905-017 Use declared type for SPARK pointer checking
A constant of pointer type is considered as mutable in SPARK,
according to SPARK RM 3.10, but this should be based on the
declared type of the constant instead of its underlying type.

The same holds for the rules regarding IN parameters of access
type, for which the declared type should be taken into account.

This ensures that various parts of GNATprove have the same interpretation
of SPARK rules, whether or not the SPARK boundary is known, and whether
or not entities have been marked for translation inside gnat2why.

* gnat2why/flow/flow_utility.adb
(Is_Variable): Use Is_Access_Type to test declared type.

* gnat2why/spark/spark_util.adb
(Is_Global_Entity): Use Is_Access_Type to test declared type.

* gnat2why/why/gnat2why-borrow_checker.adb
(Handle_Globals, Check_Call_Statement): Use declared type of
 parameter or global.

Change-Id: I4aca12027b00ee21cdc10de5eaf5de8471e05534