S903-016 Test diff after Why3 is merged
These are the diffs for the why3 merge. Regressions are present but we
argue these are minor regressions.

* testsuite/gnatprove/tests/P719-001__force/*
Its ok because the test only check that -f works. A regression on altergo
is ok on this test as seen in test.py.

* testsuite/gnatprove/tests/S117-008__pointer/*
One regression but many other unproved goal are present in this file

* testsuite/gnatprove/tests/ghc_sort/*
Raise steps to 200 to make it proved again

* testsuite/gnatprove/tests/prime_numbers/*
One regression. Others assertions are unproved in this test. Doubling
steps is not enough to restore the proved goal.

* testsuite/gnatprove/tests/tetris/*
Increasing steps solve other medium messgaes but not the one added. It
makes the test worryingly long so we keep it with the regression.

* testsuite/gnatprove/tests/R509-011__higher_order_proof/*
Regressions probably due to mailserver so not commited

* testsuite/gnatprove/tests/*
Some improvements in counterexamples. Some differences in provers.

Change-Id: Icbd9a7d9a00a743b16be4416b03c6f21caeccc7c