S905-002 Fix initial permission of declared objects
Observing objects declared of type access-to-constant should have read-only
permission only, as well as constants of a type not declared as
access-to-variable (even if the type is private and completed as as
access-to-variable type).

* gnat2why/why/gnat2why-borrow_checker.adb
(Check_Not_Borrowed, Check_Not_Observed): Move outside of Process_Path.
(Is_Read_Only): Fix definition.
(Check_Statement): Do not check Assign mode on target of assignment when
 assigning to local observer/borrower not itself observed/borrowed.

* testsuite/gnatprove/tests/S905-002__pointer_observe/p.adb
* testsuite/gnatprove/tests/S905-002__pointer_observe/p.ads
* testsuite/gnatprove/tests/S905-002__pointer_observe/test.out
* testsuite/gnatprove/tests/S905-002__pointer_observe/test.py
Add test.

Change-Id: Iecb4b6beb9b25bbd8003150f24e8b9f3d7f269a1