S415-012 pledge functions on result of traversal functions
This commit allows using pledge functions to describe the result of a
traversal function in its contract.

* gnat2why/spark/spark_definition.adb
(Mark_Param): Allow calling pledge functions on the result attribute of
 a traversal function returning an access to variable type.

* gnat2why/why/gnat2why-expr.adb
(New_Pledge_Update_From_Assignment): Updated to handle return expressions
 of traversal functions.
(Compute_Assumption_For_Borrow): Can be called on a traversal function as
 well as on a local borrower.
(Transform_Expr): Pledge function can be called on 'Result attribute.
(Transform_Pledge_Call): Adapt so that it can be called on 'Result. The
 Old attribute can be used inside the pledge to refer to the value of the
 borrowed parameter at the time of pledge.
(Transform_Statement_Or_Declaration): Update the pledge of the result on
 return of a traversal function.

* gnat2why/why/gnat2why-subprograms.adb
(Generate_VCs_For_Subprogram): If the subprogram is a traversal function
 returning a borrower, declare a reference for the pledge of its result.
(Create_Function_Decl,
 Emit_Post_Axiom): Bind the identifier for the pledge of the result in
 post of traversal functions.

* gnat2why/why/why-gen-pointers.adb
(New_Pledge_For_Call): Create the expression of the pledge of a call.
(New_Pledge_Fun_Call): Renamed to New_Pledge_Call to avoid confusion
 with the previous one.
(Result_Pledge_Id): Create an identifier for references to the pledge of
 the result of a traversal function in its postcondition.

* tests/S415-012__bad_pledge_calls/test.out
Change in error messages.

* tests/S415-012__pledge_for_traversal
New test.

Change-Id: Ia90fa28ee89f96408fb7069178cc030c254edd8b