S912-009 reject assignment into constant object in marking
In SPARK, constant objects of a deep type which is not visibly an access
type are considered to be constants (that is, we are not allowed to modify
them). Reject assignment into such objects in marking.

* gnat2why/spark/spark_definition.adb
(Mark): Reject assignment into constant objects.

* gnat2why/spark/spark_util.adb
(Is_Constant_In_SPARK): Return True if the object is a true constant in
 SPARK.

* gnat2why/why/gnat2why-borrow_checker.adb
(Is_Read_Only): Use Is_Constant_In_SPARK.

* gnat2why/why/gnat2why-util.adb
(Is_Mutable_In_Why): Use Is_Constant_In_SPARK.

* tests/RC17-022__private_pointer
Update test so that it is not rejected by SPARK rules.

* tests/S823-011__constant_pointer/test.out
The code is rejected in marking now.

* tests/S912-009__flow_constant_access
Enable test.

Change-Id: I4183cdb1304d0cc5dc34fe23342fcb5552416df0