S618-006 Forbid variable input in the expression of a borrow
For the declaration of a local borrow, we need to be able to relate the
borrowed expression and the borrower in a so-called pledge predicate,
which requires the borrowed expression not to have variable input, except
for its root object.

We first define the notion of root name in SPARK RM 3.10, before adding
this case to the SPARK RM rule 4.4(2) defining which expressions should not
have variable input.

* docs/lrm/source/declarations-and-types.rst
* docs/lrm/source/names-and-expressions.rst
Define root name based on the existing root object, and the notion of
enclosing object which we use elsewhere in SPARK RM.
Reject object declarations that define a borrowing operation when the
borrowed expression has variable input.

* gnat2why/flow/flow-analysis-sanity.adb
(Traverse_Declarations_Or_Statements): Check borrowing declarations for
 variable input in the borrowed expression.

* testsuite/gnatprove/tests/S618-006__pointer/borrow_with_variables.adb
* testsuite/gnatprove/tests/S618-006__pointer/test.out
* testsuite/gnatprove/tests/S618-006__pointer/test.py
Add test.

* testsuite/gnatprove/tests/S618-006__local_borrowers
Update test.

Change-Id: I5bc18b148f3fdc57d49afb83ad6515769e2a47ca