S730-037 proper routine from borrow-checker to flow conversion
The previously (mis)used Get_Assignment_Target_Properties rightly crashes on
qualified expressions, which cannot appear on the LHS of an assignments, but
are perfectly legal on the RHS of a local borrower.

* gnat2why/flow/flow_utility.ads
* gnat2why/flow/flow_utility.adb
(Path_To_Flow_Id): new routine for conversion between borrow checker and
 flow representation of objects.

* gnat2why/flow/flow-control_flow_graph.adb
(Do_Subprogram_Or_Block): use new routine.

* testsuite/gnatprove/tests/S730-037__pointer/
Enable, but only in flow mode (it still crashes in proof).

Change-Id: Ie2548dbb9e93158a6a5e3f516837eb841c3f18d8